requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
lightgbm==4.1.0
python-dotenv==1.0.0
googlemaps==4.10.0
openai==1.3.5
python-multipart==0.0.6
